split text.ts
